feat: Implements the parseEvent function so that it can parse a valid push event and return the corresponding CI Job. (#26)
* Modifies build.gradle to specify the JDK version to be 11
Fixes #11

* feat: Implements the parseEvent function so that it can parse a valid push event and return the corresponding CI Job.
A valid test is also present to test parsing a valid push event. Invalid and "GET" requests or other events might need to be handled next.
build.gradle is modified correspondingly so that it can successfully build with java 8 by removing the dependency of java 11 otherwise "invalid source release: 11" is throwed, jetty version is modified along with it, and necessary dependency of json, springframework added.